refactor : reorder shipping address
