+++ src/Core/Application/Aggregates/Customer/ShippingAddress/UpdateShippingAddressViewModel.cs
@@ -1,12 +1,11 @@
-﻿using Domain.Aggregates.Customers;
-using System.ComponentModel.DataAnnotations;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace Application.Aggregates.ShippingAddress
+namespace Application.Aggregates.Customer.ShippingAddress
 {
-    public class CreateShippingAddressViewModel
+    public class UpdateShippingAddressViewModel 
     {
         [Display(ResourceType = typeof(Resources.DataDictionary),
-         Name = nameof(Resources.DataDictionary.Country))]
+        Name = nameof(Resources.DataDictionary.Country))]
         public string Country { get; set; }
 
         [Display(ResourceType = typeof(Resources.DataDictionary),
@@ -27,5 +26,6 @@ namespace Application.Aggregates.ShippingAddress
 
         public Guid CustomerId { get; set; }
        
+        public Guid Id { get; set; }
     }
 }
