+++ src/Persistence/Persistence/Aggregates/ShippingAddress/ShippingAddressRepository.cs
@@ -1,5 +1,5 @@
 ﻿using Domain.Aggregates.Customers;
-using Domain.Aggregates.ShippingAddress;
+using Domain.Aggregates.Customers.ShippingAddress;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -11,22 +11,22 @@ namespace Persistence.Aggregates.ShippingAddress
 {
     public class ShippingAddressRepository(UniBazzarContext uniBazzarContext) : IShippingAddressRepository
     {
-        public void AddShippingAddress(Domain.Aggregates.ShippingAddress.ShippingAddress entity)
+        public void AddShippingAddress(Domain.Aggregates.Customers.ShippingAddress.ShippingAddress entity)
         {
             uniBazzarContext.Add(entity);
         }
 
-        public Task<List<Domain.Aggregates.ShippingAddress.ShippingAddress>> GetAllShippingAddressAsync(Guid CustomerId)
+        public Task<List<Domain.Aggregates.Customers.ShippingAddress.ShippingAddress>> GetAllShippingAddressAsync(Guid CustomerId)
         {
             return uniBazzarContext.ShippingAddresses.Where(x=> x.CustomerId == CustomerId).ToListAsync();
         }
 
-        public Task<Domain.Aggregates.ShippingAddress.ShippingAddress> GetShippingAddressAsync(Guid id)
+        public Task<Domain.Aggregates.Customers.ShippingAddress.ShippingAddress> GetShippingAddressAsync(Guid id)
         {
            return uniBazzarContext.ShippingAddresses.FirstOrDefaultAsync(x => x.Id == id);
         }
 
-        public void Remove(Domain.Aggregates.ShippingAddress.ShippingAddress entity)
+        public void Remove(Domain.Aggregates.Customers.ShippingAddress.ShippingAddress entity)
         {
             uniBazzarContext.Remove(entity);
         }
