+++ src/Presentation/Server/Infrastructure/Extentions/ServiceCollections/DomainApplicationsExtention.cs
@@ -1,16 +1,16 @@
 ﻿using Application.Aggregates.Categories;
 using Application.Aggregates.Customer;
+using Application.Aggregates.Customer.ShippingAddress;
 using Application.Aggregates.Products;
-using Application.Aggregates.ShippingAddress;
 using Application.Aggregates.Stores;
 using Application.Aggregates.Units;
 using Application.Aggregates.Users;
 using Domain;
 using Domain.Aggregates.Categories;
 using Domain.Aggregates.Customers;
+using Domain.Aggregates.Customers.ShippingAddress;
 using Domain.Aggregates.Products;
 using Domain.Aggregates.Products.ProductFeatures;
-using Domain.Aggregates.ShippingAddress;
 using Domain.Aggregates.Stores;
 using Domain.Aggregates.Units;
 using Domain.Aggregates.Users;
